requirements.txt
Django==3.1.4
django-datatables-view==1.19.1
django-extensions==3.1.0
djangorestframework==3.12.2
pyndlsearch==1.0
reportlab==3.5.56
requests==2.25.1
